move resolution info class into own module
