+++ frontend/javascripts/oxalis/model/bucket_data_handling/data_cube.ts
@@ -8,7 +8,7 @@ import type { ProgressCallback } from "libs/progress_callback";
 import { V3 } from "libs/mjs";
 import { VoxelNeighborQueue2D, VoxelNeighborQueue3D } from "oxalis/model/volumetracing/volumelayer";
 import { areBoundingBoxesOverlappingOrTouching, castForArrayType } from "libs/utils";
-import { ResolutionInfo, getMappingInfo } from "oxalis/model/accessors/dataset_accessor";
+import { getMappingInfo } from "oxalis/model/accessors/dataset_accessor";
 import { getSomeTracing } from "oxalis/model/accessors/tracing_accessor";
 import { globalPositionToBucketPosition } from "oxalis/model/helpers/position_converter";
 import { listenToStoreProperty } from "oxalis/model/helpers/listener_helpers";
@@ -24,6 +24,7 @@ import TemporalBucketManager from "oxalis/model/bucket_data_handling/temporal_bu
 import Toast from "libs/toast";
 import type { Vector3, Vector4, BoundingBoxType, LabelMasksByBucketAndW } from "oxalis/constants";
 import constants, { MappingStatusEnum } from "oxalis/constants";
+import { ResolutionInfo } from "../helpers/resolution_info";
 
 const warnAboutTooManyAllocations = _.once(() => {
   const msg =
