+++ frontend/javascripts/test/model/model_resolutions.spec.ts
@@ -1,11 +1,9 @@
 import "test/mocks/lz4";
 import test from "ava";
-import {
-  convertToDenseResolution,
-  getResolutionUnion,
-} from "oxalis/model/accessors/dataset_accessor";
+import { getResolutionUnion } from "oxalis/model/accessors/dataset_accessor";
 import { Vector3 } from "oxalis/constants";
 import { APIDataset } from "types/api_flow_types";
+import { convertToDenseResolution } from "oxalis/model/helpers/resolution_info";
 
 test("Simple convertToDenseResolution", (t) => {
   const denseResolutions = convertToDenseResolution([
